readability-implicit-bool-conversion: safe/ok bits
```shell
cd $os_root/src
run-clang-tidy -header-filter='.*' -checks='-*,readability-implicit-bool-conversion' -fix -format "$(pwd)/.*\.(cpp|hpp)"
```

+++ src/utilities/filetypes/EpwFile.cpp
@@ -1593,7 +1593,7 @@ static double stringToDouble(const std::string& string, bool* ok) {
 }
 
 Date EpwDataPoint::date() const {
-  return {MonthOfYear(m_month), m_day, m_year};
+  return {MonthOfYear(m_month), static_cast<unsigned int>(m_day), m_year};
 }
 
 void EpwDataPoint::setDate(Date date) {
